Added reset command to output node
Sometimes devices need a reset before they accept commands.
The original SW sends that reset before setting the temperature.
Not sure if it sends the reset all the time or only when required.

This change adds support for the reset command to the output node.
The underlying library supported that command anyway.

Reset can be send as a sepparate command or before a temperature command.